PLANET-8084 Refactor Cookies block to utilize block.json
This is to follow Gutenberg methods
