--- assets/src/blocks/Cookies/CookiesEditorScript.js
@@ -1,3 +1,0 @@
-import {registerCookiesBlock} from './CookiesBlock';
-
-registerCookiesBlock();
