+++ assets/src/blocks/Cookies/style.scss
@@ -1,3 +1,7 @@
+@import "../../scss/base/tokens";
+@import "../../scss/base/variables";
+@import "../../scss/base/mixins";
+
 .cookies-block {
   .custom-control-description,
   .custom-control input[type="checkbox"] ~ .custom-control-description {
@@ -47,5 +51,38 @@
 
   .cookies-checkbox-description {
     font-size: var(--font-size-xxs--font-family-tertiary);
+  }
+
+  // Editor styles
+  .cookies-checkbox-description.rich-text {
+    display: inline-block;
+  }
+
+  .field-reset-button {
+    display: inline-block;
+    color: var(--grey-600);
+    margin-inline-start: $sp-2;
+    font-family: var(--font-family-tertiary);
+    font-size: 13px;
+    vertical-align: middle;
+    white-space: nowrap;
+
+    .cta {
+      cursor: pointer;
+
+      &:hover {
+        text-decoration: underline;
+      }
+    }
+
+    .info {
+      margin-inline-end: 8px;
+      border-radius: 50%;
+      width: 18px;
+      height: 18px;
+      display: inline-block;
+      text-align: center;
+      border: solid 1px var(--grey-600);
+    }
   }
 }
