+++ assets/src/blocks/Cookies/view-script.js
@@ -1,8 +1,9 @@
 import {createRoot} from 'react-dom/client';
-import {CookiesFrontend} from './CookiesFrontend';
+import metadata from './block.json';
+import {CookiesFrontend} from './cookies';
 
 // Fallback for non migrated content. Remove after migration.
-document.querySelectorAll('[data-render="planet4-blocks/cookies"]').forEach(
+document.querySelectorAll(`[data-render="${metadata.name}"]`).forEach(
   blockNode => {
     const attributes = JSON.parse(blockNode.dataset.attributes);
     const rootElement = createRoot(blockNode);
