+++ assets/src/scss/blocks.scss
@@ -5,7 +5,6 @@
 @import "blocks/Accordion/AccordionStyle";
 @import "blocks/CarouselHeader/CarouselHeaderStyle";
 @import "blocks/Columns/ColumnsStyle";
-@import "blocks/Cookies/CookiesStyle";
 @import "blocks/Counter/CounterStyle";
 @import "blocks/ActionsList/ActionsListStyle";
 @import "blocks/PostsList/PostsListStyle";
