+++ assets/src/scss/editorStyle.scss
@@ -37,7 +37,6 @@
 @import "blocks/Accordion/AccordionEditorStyle";
 @import "blocks/CarouselHeader/CarouselHeaderEditorStyle";
 @import "blocks/Columns/ColumnsEditorStyle";
-@import "blocks/Cookies/CookiesEditorStyle";
 @import "blocks/HappyPoint/HappyPointEditorStyle";
 @import "blocks/TableOfContents/TableOfContentsEditorStyle";
 @import "blocks/TakeActionBoxout/TakeActionBoxoutEditorStyle";
